requirements.txt
Django==3.1.3
django-cors-headers==3.5.0
django-mysql==3.9
djangorestframework==3.12.2
pika==1.1.0
mysqlclient==2.0.1
django-rest-swagger
python-dotenv==1.0.0